requirements.txt
apturl==0.5.2
asgiref==3.8.1
attrs==19.3.0
backports.zoneinfo==0.2.1
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
django==4.2.27
djangorestframework==3.15.2
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
exceptiongroup==1.3.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
iniconfig==2.1.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==26.0
Pillow==7.0.0
pluggy==1.5.0
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pyrsistent==0.15.5
pytest==8.3.5
pytest-django==4.11.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.5.5
systemd-python==234
texttable==1.6.2
tomli==2.4.0
typing-extensions==4.13.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
websocket-client==0.53.0
xkit==0.0.0
zipp==1.0.0
